Add #request method for Rails controllers
This is a first step in filling out the ActionController sigs for Rails.
Technically the `#request` method is nilable. In practice I don't
think this is intended to be the case (at least the docs indicate that
a request object is returned). Marking the result as non-nilable means
that it is much easier to use this method without having to use `T.must`
throughout a Rails codebase.