Run tests for each gem + version in isolation
As the test harness is currently setup, if one version of a gem has a
signature that conflicts with another version of the same gem (or any
other gem), then the tests will never succeed because the two versions
will conflict with each other. This change resolves these issues by
running each gem and version combination (along with the respective
`all` directory for that gem) in isolation. A failure to type-check any
one gem + version combination will fail the test harness script.