Add interface for Rails view contexts
Controllers in Rails have this method called `view_context` that
returns an instance of a view context class. The basic view context
is `ActionViews::Base`, but this can be configured with the
`view_context_class`. The requirements for a view context class are that
it can be initialized with three arguments, and that it responds to the
`render` method with a `String`.

This change adds an interface for view contexts. A lot of the argument
types are currently untyped. I believe they can be filled in later. The
main thing this change provides is the proper interface for a view
context, and the actual method that returns a view context so that Rails
applications can type-check against it.